+++ Tasks/UnityActivateLicense/UnityActivateLicenseV1/post-unity-activate-license.ts
@@ -11,7 +11,6 @@ import {
     tempDirectoryInputVariableName,
     passwordInputVariableName,
     unityEditorsPathModeInputVariableName,
-    unityProjectPathInputVariableName,
     usernameInputVariableName,
     versionInputVariableName,
     versionSelectionModeVariableName
@@ -33,7 +32,6 @@ function run() {
         const unityEditorsPath = UnityPathTools.getUnityEditorsPath(
             tl.getInput(unityEditorsPathModeInputVariableName, true)!,
             tl.getInput(customUnityEditorsPathInputVariableName));
-        const unityProjectPath = tl.getPathInput(unityProjectPathInputVariableName) ?? '';
 
         let unityVersion: UnityVersionInfoResult;
         if (versionSelectionMode === 'specify') {
@@ -63,12 +61,12 @@ function run() {
             const unityCmd = tl.tool(unityExecutablePath)
                 .arg('-batchmode')
                 .arg('-nographics')
-                .arg('-projectPath').arg(unityProjectPath)
                 .arg('-username').arg(username)
                 .arg('-password').arg(password)
                 .arg('-returnlicense')
                 .arg('-logfile').arg(logFilePath)
                 .arg('-quit');
+
             const result = unityCmd.execSync();
 
             // Unity process has finished. Set task result.