+++ Tasks/UnityActivateLicense/UnityActivateLicenseV1/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.2]
+
+### Changed
+
+- Aligned configuration options grouping and naming with other tasks
+- Updated and cleaned up external dependencies
+- Task deprecated, please use UnitySetupV2 instead for a streamlined Unity setup and configuration experience
+- [Update node version](https://github.com/Dinomite-Studios/unity-azure-pipelines-tasks/pull/259)
+
 ## [1.2.1]
 
 ### Changed